Partial upload script update
The site probably won't live in the same place on the new server. It
might not even be accessible via SSH at sopel.chat... But this is almost
definitely how the site will need to be built and what folder will need
to be copied, so this is still useful to commit.