Distraction pays off: the "migrated" box looks really good now
I thought the box alerting users to the fact that an article was
migrated from the wiki needed something, like a texture. So I went down
the CSS rabbit hole for a while and came up with something that looks
pretty darn awesome (with the help of a pattern generator)!