Migration notice should only appear if wanted
I anticipate that the "source" property will remain in the articles'
front matter even after the "migrated" property is set to false or
removed. It's good to keep that kind of metadata around, just in case it
proves to be useful for something later. But once pages are cleaned up
and the "migrated" flag turned off, they don't need to show a notice.