Add scaffolding for "Usage" articles (next to migrate)
Sopel's old wiki has a fair spread of content. The more user-targeted
pages should go here, instead of in the tutorials, to segregate the
things that require writing code from the things that don't.

The installation tutorial is a special case… It might get moved around
as the site structure evolves.