Brought over the wiki's module-documentation script
Updated and renamed for Sopel instead of Willie, and extended to take in
the path to Sopel's code/repo in case it's not stored in the expected
(relative) location.

The only thing that doesn't work at present is the module-config
parsing. That will need some extra work. There's no reason to update
_usage/module-configuration.md until then, since the script presently
just puts "This module has configuration options that are not
documented. Go bludgeon the author." for every module with a configure()
method, which is even less useful than the current outdated page.

Apparently the commands page wasn't committed due to a .gitignore rule I
didn't know about, so this "initial" version is the one generated by the
newly polished auto-doc script. That means it doesn't need a "migrated"
flag in its front matter. Yay!