"Fixed" the module configuration autodoc generator
Actually, it was "fixed" by adding back the Markdown tables that were
removed in the various modules back when things were converted to use
the new static config system.

There's already too much duplication of option documentation. Each
module has to doc its options in three different places: in its
StaticSection subclass, in its configure() method's prompts, and in the
configure() method's docstring (for this autodoc script).

I'm convinced that there must be a way to get that down to two, at most,
by somehow reading the docstrings from each member of the StaticSection
subclass defined by each module. Just need to figure out what that way
is, and implement it in code. But this will do until then.