re-ignore Gemfile.lock for now
Committing it, as recommended, makes Netlify deployment break because
their Ruby/Bundler/something is outdated (honestly I don't know exactly
what the problem there is).

So, we're going to live dangerously for a while and not pin versions of
anything with a lockfile. (It's not like we're managing all dependencies
properly, anyway. Python dependencies for building Sopel's Sphinx docs
are installed ad hoc from the build script...)