Move "migrated" notice icon to top on narrow screens
When the aside/box gets *too* skinny, the icon really doesn't work on
the side. Moving it to the top in this case with a @media query gives
that little bit more polish to the site on mobile devices.

Of course, anyone doing anything IRC-related on a mobile device is
insane, even in 2019, but still... it should at least look OK!