First pass at modernizing the navbar
We don't need a semantically meaningless "first" class any more, because
browser support for :not() and :first-child/:last-child is pretty
widespread at this point. Even older (MUCH older) browsers that don't
support it will simply get an extra dividing line that shouldn't be
there, so that's no big deal.

Switching to flexbox is the thing that might actually break something,
but I'm hoping that it'll make the navbar more elegant on smaller
screens. (Flexbox is really good at distributing items across multiple
lines if you let it, as this layout does.)