Convert FAQ to a more semantically appropriate <dl>
Using a <ul> looked OK, but it doesn't really mean anything. Giving
<li>s the classname "q" or "a" is only so meaningful.

A definition list (<dl>) is much more appropriate for this content. Each
question is a term (<dt>), and each answer is a definition (<dd>).

Of course, none of this gets around how annoying it is that Markdown
simply refuses to work inside any block-level HTML element. Oh well.