Annoyance prevention for <pre> tags in content
On narrow screens, <pre> blocks could cause the entire page layout to
shrink if the content of the <pre> was wide enough.

In addition to giving the <pre> blocks some padding (room to breathe)
just so they look nicer, telling the browser it can scroll x-axis
overflow prevents the layout breaking as above.