Initial web-based changelog generation
Generates Markdown -> HTML changelog pages from the _sopel submodule
already used for auto-documenting modules and generating Sphinx docs.

Sopel's NEWS file isn't really Markdown-compliant. It's close enough
that none of the pages I spot-checked are *broken* per se, but things
like bare links don't work because of Kramdown (they MUST be wrapped in
<angle brackets> to be auto-linked).

The best solution to this is probably going back and revising Sopel's
NEWS file, since it's inconsistent in a lot of little ways, anyway.
Ultimately, Sopel's README should become Markdown instead of rST,
too, so we're not dealing with too many different documentation
formats. Both files could be revised at the same time, in fact.

As a bonus, this changeset removes a problematic CSS style that added
a top margin to first-child elements within the .content frame. It was
quite a puzzle tracking that down, because only the first item in the
version list was misaligned... Whatever past version of myself thought
that style was a good idea, I should go back in time and smack him.