Update site documentation/instructions
UPDATING.md now covers the changelog section as well, and gives better
instructions for things in general.

Text now reflects that generated pages don't need to be updated by a
human, but will be kept up-to-date automatically based on the _sopel
submodule commit included in the repo.

The readme more accurately reflects the current home of the site
(Netlify, instead of a webserver somewhere).