Generate latest.json file automatically
I forgot to update it for three releases in a row, which will not do.

The release notes now also point to the sopel.chat changelog page for
that version, instead of a GitHub release page.

Renamed generate_changelogs.py to document_versions.py because it's
doing more than just changelog generation now. Updated all the scripts
and documentation files that point to it. (There's also a reference in
one of Sopel's documentation files that needs updating.)