Constrain prev/next link widths to avoid overlap
Noticed some ugly overlapping on mobile in portrait mode, because of the
"Up a level" links being absolutely positioned for *reasons*.

Wanted to get rid of the absolute positioning with some fancy CSS
selectors, but apparently such a thing (selecting an element with a
certain number of children) doesn't exist. This will do instead.