Merge download page into installation guide
We're already kind of duplicating a whole section of Sopel's README by
having this install guide on the site. The download page was yet another
place that said basically the same thing in a slightly different way.

Took the stuff from there that was missing, put it in the installation
page, and deleted the rest.