document_sopel_modules: guard against dir() silliness
For some reason, the list resulting from dir(module) includes names that
that module imported from elsewhere.

That means that a Sopel module which does `from sopel import module`
will have a member named `module` which passes the `hasattr('commands')`
test used by document_module() to see if it should try to document that
"function", even if it's not a function. (`sopel.module` contains a
decorator function named 'commands'.)

Checking if `func` is callable before seeing if it has an attribute
named 'commands' avoids trying to process modules as if they are
functions, and thus prevents the script from dying with a TypeError.