Abort build script on non-zero exit codes
Spurred by build errors that silently allowed the deploy to proceed
anyway with a broken page caused by a Python exception in one of the
auto-doc scripts.

Python should return a non-zero exit status if an exception is thrown,
and we definitely want the build script to abort if that happens so
Netlify will not continue the deploy and possibly break the live site.