installing: Skip listing upgrade guides without version range
With this change, it's easier to draft an upgrade guide without exposing
it on the public site. Omitting either (or both) of the front matter
fields defining the applicable version range will cause the "Installing"
page to omit that guide from the list.

It's also good for safety, since empty fields are treated as 0 by the
Liquid code that generates the list, meaning a guide with a missing
field may incorrectly show "Willie" instead of "Sopel", and that with a
missing version number. It's safer to just leave out entries missing
required field data (and it's impossible to throw an error from Liquid).