style: fix command table display on mobile
Google Webmaster Tools complained that the Commands page wasn't
mobile-friendly. In fact, it looked pretty awful in the example
screenshot, because of how wide the table is.

This is an alternative to the (admittedly simple, but kind of ugly)
option of using `overflow-x: auto` (or `scroll`).

The site doesn't use many tables, so hopefully it's safe to assume that
bolding the first column of any table will be a decent enough substitute
for having the table headers visible.