Remove GitHub banner
It screws things up on mobile because the image's square tap zone
invisibly covers part of the navigation links, and it's unnecessary
because we have a "Source Code" link in there.

After playing with pushing the banner's tap zone "behind" the navbar, it
seemed easier to just remove it. z-index behavior is not worth fussing
with when the thing's redundant anyway.