usage/cli: Redirect to Sphinx CLI documentation
Keeping the page allows it to stay in the "Usage" list as a shortcut.

Includes the same path in the front-matter and body as a hedge against
the Sphinx section file layout changing, since `{% link path/to/file %}`
will fail the build if `file` doesn't exist.

Putting a link to the target page is also a good hedge against breaking
the redirect functionality (e.g. switching away from Netlify), since
there will still be a way for users to get to where they need to be.