gitignore some extraneous files
With the Netlify build script becoming the only sane way to prepare a
local copy of the site for iterative preview using `bundle exec jekyll
serve`, the files it spits out everywhere almost necessitate a Makefile.
Until I can be bothered to create a Makefile (and `make clean` option),
ignoring the build products will do in a pinch.