tutorial, part 1: fix spelling, and model our "standard" import style
The typo is obvious.

Less obvious is that we want to encourage the `from X import Y` style of
importing things like `sopel.module` instead of importing the full
dotted name. I'd argue that anyone who will be confused by the prose
following that code block shouldn't be writing plugins yet, and needs to
start with a Python primer.