Remove outdated tutorials in favor of new plugin docs
I've tried to make sure that we won't see broken links by redirecting
any traffic that asks for one of the removed tutorial pages into the
"Developer Overview" of plugins in /docs/plugin.html and the rest of
that section in /docs/plugin/.

The Appendix page about formatting now redirects to the "Do it with
style" section of how to have plugins make the bot talk.

All internal links within the Jekyll site should be updated.