Independent generation of `latest.json` file
The new script loads releases from PyPI, independently of the changelog
generation script that used to handle this. If there is a prerelease
newer than the current stable version, that will be the `unstable`
version, instead of an alias to the current stable version.

Leans on `packaging.version` to do the heavy lifting (parsing version
specifiers into sortable `Version` objects). Fortunately we already get
`packaging` and `requests` for free when installing `_sopel` to document
it and its plugins.

Co-authored-by: James <snoopjedi@gmail.com>