usage: redirect system requirements to Sphinx installation page
The system requirements page is hella outdated. Rather than waste time
rewriting it, let's just direct people to the up-to-date requirements
section pulled in from Sopel itself, where the docs are maintained in
parallel with the actual code.

(My urge to fully migrate the WHOLE site into Sphinx and ditch the Ruby
stuff is rising…)